requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
et_xmlfile==2.0.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
httplib2==0.31.2
idna==3.11
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
playwright==1.58.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
